qpid-proton: connect_config_test: adapt for testing PKCS#11 URIs
All tests hardcode paths to PEM files. For easily testing PKCS#11 usage
for client certificates on the target, it's convenient to be able to override
the file paths without recompilation. Enable doing that by consulting a
set of new environment variables (TEST_CERT, TEST_KEY, TEST_CA,
TEST_VERIFY) and falling back to the hardcoded paths for the undefined
variables.